+++ src/components/Sidebar/index.tsx
@@ -1,6 +1,6 @@
 import { DraggablePanel } from '@/components'
 import { useResponsive } from 'antd-style'
-import React, { useState } from 'react'
+import React, {useEffect, useState} from 'react'
 import styled from 'styled-components'
 import PromptGroup from './PromptGroup'
 
@@ -62,6 +62,11 @@ interface SidebarProps {
 const Sidebar: React.FC<SidebarProps> = ({ children, loading }) => {
   const { mobile } = useResponsive()
   const [expand, setExpand] = useState<boolean>(!mobile)
+
+  useEffect(() => {
+    setExpand(!mobile)
+  },[mobile])
+
   return (
     <DraggablePanel placement="left" defaultSize={{ width: 280 }} isExpand={expand} onExpandChange={setExpand}>
       <SidebarView>