+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.4.1](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/compare/v1.4.0...v1.4.1) (2023-04-21)
+
+
+### 🐛 Bug Fixes
+
+* dragpanel auto hide in mobile view, fixed ([a7e5d47](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/commit/a7e5d47)), closes [#35](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/issues/35)
+
+# [1.4.0](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/compare/v1.3.2...v1.4.0) (2023-04-21)
+
+
+### ✨ Features
+
+* add suggestions to prompt editor ([bebebe4](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/commit/bebebe4))
+
 ## [1.3.2](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/compare/v1.3.1...v1.3.2) (2023-04-21)
 
 
@@ -9,10 +23,9 @@
 
 ## [1.3.1](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/compare/v1.3.0...v1.3.1) (2023-04-21)
 
-
 ### 🐛 Bug Fixes
 
-* send func, tags lost ([0be6f28](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/commit/0be6f28))
+- send func, tags lost ([0be6f28](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/commit/0be6f28))
 
 # [1.3.0](https://github.com/canisminor1990/sd-web-ui-kitchen-theme/compare/v1.2.2...v1.3.0) (2023-04-20)